Moved jest extensions to its own file
